Implement - merge JSON into existing sound pack
The code in this commit combines the generated-sounds.json file that it
creates earlier with the sounds.json file in the existing sound pack
folder, and overwrites the sounds.json file with the combination.

During the process, I found that using dict.append() caused the sound
dictionaries to become alphabetized, which I didn't want, because I'm
really used to seeing "name" first, then volume, then anything else.
I ended-up using OrderedDict to keep keys in the order I add them.

My intention is to refactor this mess, but it works for now.

Note: this is my first commit message since discovering the 7 rules for
good commit messages. Try not to judge my other commit messages too
harshly.  :)