Refactor JSON generation code
Moved code from Main() into get_sound_name_start_index() function, to
encapsulate the weird calculation I have for determining where to break
the path name into something that works for Minecraft sounds.json

Changed the name of get_combined_json() to get_combined_events(),
because it's actually returning a dictionary of events and not
technically JSON.

Moved code from Main() into get_generated_events() to de-clutter Main()
a bit.  Now I might have a testable function on my hands.

Renamed the regular expression for testing the suitability of a path
to be used in sounds.json from "good_name" (which isn't) to
mc_naming_rules, which better describes its purpose.

Renamed "source" and "sorted_events" in Main() to "source_path" and
"generated_events" respectively for clarity.

Fixed a logging bug; the final display to the user showed only the
generated JSON and not the combination

Eliminated a few debugging print statements that I hadn't used
in a long time.