Ignore folders with no .ogg files in them
Added an if statement to get_generated_events() to skip processing the
current file if the folder in question has no .ogg files in it.

This commit includes a new test folder structure with dummy .ogg files
and a few .subtitles files as well. It covers every situation I could
think of: empty folders, folders with only .subtitles files, folders
with only .ogg files and folders with both .ogg and .subtitles.

While I was in there, I also got rid of a squiggly in my header comment
and added code to resolve the incoming paths so that the script doesn't
report to the user: "I did a thing in folder ./"

Fixes issue #1