Skip non .ogg files, and folders with no .ogg files in any subdirectory
This was an adventure, and extremely frustrating.

My first issue was that I couldn't use any print statements on the
original factory, because it relied heavily on generators, which were
losing their values the moment I tried to show myself what they were.

My second issue was that I was trying to keep the generic reusability
of the original factory design; namely that you could pass in a list of
suffixes, and it would work for any use-case. This was adding complexity
that I did not need for my purposes and just confusing the heck out of
me.

My actual issue was the use of the .update() function to add
file/folder names to the ignored list because, being relatively new to
Python, I was not aware that .update() would add a string as a series of
individual letters rather than whole strings. Once I switched to .add()
instead, I was able to actually ignore things.

So now the code is debuggable, it's more readable, and it actually
ignores the things I want it to ignore. Huzzah!

Resolves issue #4