Show the user a list of name collisions and allow graceful exit
Added code to check all the incoming files against files in the target
location. If any of the incoming files has the exact same name and
relative path as any of the target files, that file path is added to a
list and shown to the user in the console. The user can then decide
whether to exit and correct the name, or proceed with the overwrite.

Implements issue #5