Prevent error when not specifying a target folder
The script allows a user to not specify a target folder if they don't
care about copying files. This worked fine until I made a change to
the handle_command_line() function in an attempt to resolve relative
paths so that the user didn't get messages that say "file created in
folder . "

From the looks of it, resolving the paths in the location that I chose
caused them to no longer be "NoneType" but rather "None", which meant
that trying to resolve them down the line caused a crash.

This is a quick fix at best, and probably shouldn't be trusted. This
issue just reminds me how useful unit tests are, and I need to get
some of those into the project as quickly as possible.

Resolves issue #14 (for now)