Prevent question to copy files when there are no files to copy
While I was in there, I slightly altered the printout that the user
sees because exiting when the json is empty caused a lack of messages
that just seemed confusing, so it was time to refactor the output as
well.  I added a source or target below the appropriate banner, and
colored it in cyan to make it stand out.

A lot of the impetus behind the extra changes in this commit were due
to the fact that I've created a Fish function that runs this application
three times in succession, and with some of the messages missing, it was
difficult to see which runs were being processed at any given step, and
which runs were being skipped.

Resolves issue #15