Refactor get_combined_events() to make it more testable
When starting work on issue #9, I noticed that get_combined_events()
depended on another function that read data from the file system. It
seemed pretty obvious to do that outside the function and pass in the
sound events instead of a path. That way it can be tested without
having to mock the file system.

Also added type hinting to the function definition.

While I was working on this, I discovered that a null target path can
still crash the application, so I modified line 456 to check for that
before trying to resolve the path. It's possible that I don't even
need to resolve the path, since it should have been resolved already,
but for now, it fixes that issue.

Contributes to issue #9