Add a test that makes sure the event names are being sorted correctly
In the process of creating this test, I discovered that I needed two
events that were different from the two in the existing
two_sounds_in_different_events fixture, so instead of creating yet
another fixture, I decided to make a change to the one that already
existed.

While I was doing that, I decided to make some of the .ogg file names
in the sounds lists slightly more logical.  I'm still not entirely
happy with it, but I feel like it's better than it was.

Coverage is still only 22%, but I don't think I'm done with this
function yet.

Contributes to issue #9