Add three new tests to verify that subtitle is processed correctly
A subtitle can be imported in the following conditions:

    1. If there is not already a subtitle in the sound event
    2. If the incoming subtitle is not missing - another extreme edge
       case that would probably never happen if I consistently use this
       application.  On the other hand, real life is messy

The application should _not_ import a subtitle if one already exists.
I'm assuming that the one in sounds.json is correct, which may or may
not be the case, but I need to assume some kind of baseline.

Coverage is still at 22%, but we're uncovering issues that could exist
and we haven't accidentally stumbled upon them yet.

Contributes to issue #9