Add a new class to handle default values for sound event properties
The reason I'm doing all this is because it solves a couple of problems
that make the get_generated_events() function harder to test. Once I
get this object working and tested, I should be able to massively
refactor get_generated_events().

Moved TypedDict classes to their own file, to better facilitate
importing them in multiple places.

Created a new "Defaults" class, that will eventually serve-up all the
default values defined in a json file that I haven't created yet.  This
class needs a lot of work, and it's late and I'm going to bed, so this
is a safety check-in to make sure I don't forget to do it.

Created several tests for the get_sound_event() function on the Defaults
class.  At this point, I think I've tested everything that I can, and
I'll need to start on the get_sound() function after I sleep.

Contributes to issue #9