Add 29 new tests to the Defaults class
Testing the get_sound function. Every possible output option is being
tested now. If an event default doesn't exist for a particular setting,
it should use the "all" setting, which is a catch-all for every event
for which a default hasn't been specified.

Also testing what it looks like to print the values as a string. I'm a
little concerned that it uses single-quotes rather than double-quotes,
but according to a Stack Overflow question, using json.dumps should
correct this. On the other hand, when would I use the string
representation in the program and show it to a user? Test coverage, I
guess...

defaults.py has 100% test coverage, but I'm still thinking of adding a
few more tests.  For example, I should have validation for the incoming
values.  Right now, you can put in any string or any float, and this
would not necessarily work for Minecraft.

Overall coverage is at 69%, but I'm considering the 50% goal to only
apply to sound_pack_indexer.py, which is now at 17% ever since I moved
the TypedDict objects to a separate file.

Contributes to issue #9