Add defaults.json file
This file contains all the default values I've been using so far
(provided I spelled them all correctly.)  Once implemented, it should
eliminate the .subtitles files entirely (and all the complexity that
was added due to their inclusion.)

The idea is that this file should be copied to the namespace folder in
your staging area, and it would provide all the default values in one
location so that the program logic could avoid having to wrestle with
special files. It's all "ogg" from now on.

Contributes to issue #9