Refactor the event generation process
There are now two functions:

    process_ogg_files() takes a list of .ogg files and passes or rejects
    based on Minecraft's regex for valid path names. It also trims the
    path prior to the first folder under the "sounds" folder. It returns
    these results (including warnings), and the list of sound files that
    is returned is then passed into get_generated_events().

    get_generated_events() now uses the new Defaults object to build
    the events. A list of trimmed and filtered sound files is passed in,
    and all of the custom .subtitles and TinyTag nonsense has been
    stripped out, making it much more likely to unit-test.

This is a safety commit. I haven't finalized tests for this structure
yet, but an initial run generated a json file just fine, from what I
could tell. (Imagine this used to be the only way I did testing ... and
I trusted it!)

Contributes to issue #9