Change the way file collisions are being checked (again!)
In order to test the function check_for_overwritten_files(), I moved
the gathering of .ogg files from the target back to main(). I am now
passing the list into check_for_overwritten_files() instead of creating
it within the function. This allowed me to avoid mocking the filesystem
so that I could write tests for this code.

Code coverage is now at 41%

Contributes to issue #9