Refactor handle_command_line to accept functions as arguments
Now that the function takes functions as arguments, I can stub those
functions in tests using pytest fixtures, and avoid having to mock the
filesystem, making the handle_command_line() function testable.

Code coverage is now 55%, with a total of 72% over the three code files
that I'm actually responsible for. I'll be testing this in production
over the weekend, and when I'm satisfied that it's fully functional,
I'll be changing the version number to 1.0 and closing issue #9

Contributes to issue #9