Add SoundEventCatalog and tests to verify its one method
The data for SoundEventCatalog is built using a script that crawls
through the minecraft vanilla sounds.json looking for sound event
names. It pulls the first segment of each name (the word prior to the
first dot), and makes that a key to a dictionary. The value(s) pointed
to by that key are a list of sound event names that all begin with that
same first segment.

The technique used in get_sound_event_name is to look for the first
occurrence of the dictionary key in an incoming path, and then build a
sound event name from the path starting at the index of the match.

If the event name constructed from the path doesn't match any of the
names in the catalog, a SoundEventValueError is raised. If it walks
through the entire path and doesn't match the dictionary key at all,
another SoundEventValueError is raised with a different message.

This should replace get_sound_event_name from the main application, as
it has the ability to verify a constructed name against a list of names
that Minecraft actually expects.

Implements issue #17