Refactor get_generated_events to use the new SoundEventCatalog object
Eliminated the old get_event_name function and replaced it with the
get_sound_event_name method on the SoundEventCatalog object instead.
Had to re-work a ton of tests.  Also, all the tests for the old
get_event_name function had to be thrown away.

get_generated_events is now returning warnings, due to needing to catch
exceptions raised by get_sound_event_name, so it needed a handful more
tests due to that.

I guess this contributes to issue #9