Rename project to spindex and set version to 1.0
Since we have 58% coverage on the main file alone, and 75% overall, it's
time to make this a 1.0 version. In the process, I renamed the project
folder and the main file to "spindex" so that it's easier to get to the
unique letter between this project and "sound-pack-checker" - that has
been rather annoying for months. Also followed pep8 and shortened every
line that went past the recommended column due to PyCharm defaulting to
120 for some reason.

There was also a strange set of include statements that were trying to
include the Typed Dictionaries from the main .py file, where they
haven't lived for a long time, yet somehow the tests ran right after I
changed the name of the project folder and rebuilt the project. I'm not
sure what caused it to notify me of the problem, but that's fixed now, I
guess...

Satisfies issue #9