Refactor path validation logic
While writing tests for these functions, I discovered that
validate_path_architecture was only relevant to the source path, since
the target path function was creating the proper structure if it didn't
exist, so checking it afterwards is redundant.

Moved the call to validate_target to after the user chooses whether to
even have the application do the transfer, so that if it is aborted at
that stage, we don't even need to create the directory structure at all.

The functions now raise exceptions (including a custom exception) rather
than returning a string. I'm not sure if that's the right way to go, but
it seems better than checking for None otherwise printing a string.

Added unit tests for as much as I could think of.